Move two scheduling structures to better locations
