+++ qiskit/scheduler/schedule_circuit.py
@@ -19,7 +19,7 @@ from qiskit.circuit.quantumcircuit import QuantumCircuit
 from qiskit.exceptions import QiskitError
 from qiskit.pulse.schedule import Schedule
 
-from qiskit.scheduler.models import ScheduleConfig
+from qiskit.scheduler.config import ScheduleConfig
 from qiskit.scheduler.methods.basic import as_soon_as_possible, as_late_as_possible
